1. Merge the below test cases into one case named VERIFY-LINUX-CONFIGURATION.     BVT-IS-ROOT-PASSWORD-DELETED     BVT-VERIFY-SSHD-CONFIG     BVT-LAST-CONSOLE-CHECK     BVT-VERIFY-BASH-HISTORY 2. Merge the below test cases into one case named VERIFY-LINUX-DISK-SETUP.     BVT-ROOT-DEVICE-TIMEOUT-CHECK     BVT-SWAP-CHECK     BVT-MTAB-ENTRY-CHECK     BVT-VERIFY-UUID-FSTAB
