--- Documents/Azure BVT automation/BVT-IS-ROOT-PASSWORD-DELETED.txt
@@ -1,32 +1,0 @@
-Login on your Hyper-V server.
-
-Start PowerShell with the Run As Administrator option.
-
-Deploy VM's in Azure.
-	1.	Create a hosted service.
-	2. 	Deploy a Virtual Machine in newly created hosted service.
-	
-Execute Test :
-	1.	Login to newly created VM.
-	2.	Get waagent logs. (Execute-"cat /var/log/waagent.log" on the VM)
-	3.	Verify that output contains-"Root password deleted."
-
-Cleanup the test VMs.
-	1.	If test is PASS, delete all VMs under test. 
-	2.	If test is FAIL, collect the logs and preserve deployed VMs.
-
-Edit the XML file used to drive the test
-    <test>
-		<testName>BVT-IS-ROOT-PASSWORD-DELETED</testName>
-		<testScript>BVT-IS-ROOT-PASSWORD-DELETED.py</testScript>
-		<testScriptPs1>BVT-IS-ROOT-PASSWORD-DELETED.ps1</testScriptPs1>
-		<files>remote-scripts\BVT-IS-ROOT-PASSWORD-DELETED.py,remote-scripts\azuremodules.py</files>
-		<setupType>SingleVM</setupType>
-		<TestType></TestType>
-		<TestFeature></TestFeature>
-    </test>
-	
-Run the tests defined in the .xml file.
-           From the PowerShell command line, issue the following command:
-           Run all BVT tests:
-.\AzureAutomationManager.ps1 -xmlConfigFile azure_ica_all.xml -runtests -Distro CentOS -cycleName BVT
